Add soundcloud sync support

+++ fileio.js
@@ -1,4 +1,5 @@
 import fs from 'fs';
+import { exit } from 'process';
 
 export const getPath = (libraryPath, path = '') => {
 	const basePath = libraryPath.replace(/\\/g, '/').replace(/\/$/, '');
@@ -8,24 +9,39 @@ export const getPath = (libraryPath, path = '') => {
 
 const DBName = 'db.json';
 
+const emptyDB = { version: 1, entries: [] };
+
 export const getDB = (libraryPath) => {
 	const DBPath = getPath(libraryPath, DBName);
 	if (!fs.existsSync(DBPath)) {
 		console.log('DB not found, creating new one');
-		fs.writeFileSync(DBPath, '[]');
+		fs.writeFileSync(DBPath, JSON.stringify(emptyDB(), null, 4));
 		return [];
 	}
+	let db = null;
 	try {
-		return JSON.parse(fs.readFileSync(DBPath));
+		db = JSON.parse(fs.readFileSync(DBPath));
 	} catch (e) {
 		console.log('Error reading DB', e);
-		return {};
+		exit(1);
 	}
+	const version = getDBVersion(db);
+	if (version === 0) {
+		db = migrateDB0to1(db);
+		saveDB(libraryPath, db.entries);
+	} else if (version !== 1) {
+		console.log('DB version not supported, please update the repo');
+		exit(1);
+	}
+	return db.entries;
 }
 
 export const saveDB = (libraryPath, db) => {
 	const DBPath = getPath(libraryPath, DBName);
-	fs.writeFileSync(DBPath, JSON.stringify(db, null, 4));
+	fs.writeFileSync(DBPath, JSON.stringify({
+		version: 1,
+		entries: db,
+	}, null, 4));
 }
 
 export const getLibraryMusics = (libraryPath) => {
@@ -42,13 +58,28 @@ export const getTmpPath = (libraryPath) => {
 	return getPath(libraryPath, 'tmp');
 }
 
-export const clearTmp = (libraryPath) => {
+export const clearTmp = (libraryPath, times = 50) => {
+	if (times === 0) {
+		console.log('Failed to clear tmp folder');
+		return;
+	}
 	const tmpPath = getTmpPath(libraryPath);
-	fs.rmSync(tmpPath, { recursive: true });
+	try {
+		fs.rmSync(tmpPath, { recursive: true });
+	} catch (e) {
+		if (e.code !== 'ENOTEMPTY') {
+			setTimeout(() => {
+				clearTmp(libraryPath, times - 1);
+			}, 100);
+		}
+		throw e;
+	}
+	
 }
 
 export const formatFileName = (fileName) => {
 	fileName = fileName.replace(/[<>:"\/\\|?*]/g, '');
+	fileName = fileName.replace(/(\u00a9|\u00ae|[\u2000-\u3300]|\ud83c[\ud000-\udfff]|\ud83d[\ud000-\udfff]|\ud83e[\ud000-\udfff])/g, ''); // emojis
 	fileName = fileName.trim().replace(/\.+$/, '');
 	if (fileName.length === 0) {
 		return 'invalid_fileName';
@@ -57,4 +88,21 @@ export const formatFileName = (fileName) => {
 		fileName = fileName.substring(0, 255);
 	}
 	return fileName;
+}
+
+const getDBVersion = (db) => {
+	return db?.version ?? 0;
+}
+
+const migrateDB0to1 = (db) => {
+	console.log('Migrating DB from version 0 to 1...');
+	return {
+		version: 1,
+		entries: db.map(entry => {
+			return {
+				source: 'youtube',
+				...entry,
+			}
+		})
+	};
 }
